funciona el findFilmsById y el GetAllFilms con filtros de title y genreName

+++ src/main/java/com/jorgea/PFC/to/PageResponseTo.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 public class PageResponseTo<T> {
     private List<T> content;
-    private int currentPage;
+    private long currentPage;
     private int totalPages;
 
-    public PageResponseTo(List<T> content, int currentPage, int totalPages) {
+    public PageResponseTo(List<T> content, long currentPage, int totalPages) {
         this.content = content;
         this.currentPage = currentPage;
         this.totalPages = totalPages;
@@ -21,11 +21,11 @@ public class PageResponseTo<T> {
         this.content = content;
     }
 
-    public int getCurrentPage() {
+    public long getCurrentPage() {
         return currentPage;
     }
 
-    public void setCurrentPage(int currentPage) {
+    public void setCurrentPage(long currentPage) {
         this.currentPage = currentPage;
     }
